+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0-alpha.2] - 2025-04-02
+
 ### Added
 
 - Add support for a few missing R5900EE instructions.
@@ -702,7 +704,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First version
 
-[unreleased]: https://github.com/Decompollaborate/rabbitizer/compare/master...develop
+[unreleased]: https://github.com/Decompollaborate/rabbitizer/compare/2.0.0-alpha.2...🦀
+[2.0.0-alpha.2]: https://github.com/Decompollaborate/rabbitizer/compare/2.0.0-alpha.1...2.0.0-alpha.2
 [2.0.0-alpha.1]: https://github.com/Decompollaborate/rabbitizer/compare/1.12.0...2.0.0-alpha.1
 [1.12.0]: https://github.com/Decompollaborate/rabbitizer/compare/1.11.2...1.12.0
 [1.11.2]: https://github.com/Decompollaborate/rabbitizer/compare/1.11.1...1.11.2